Refactor LoRA variant classification and merging
